backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
geoalchemy2==0.14.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.26.0
firebase-admin==6.4.0
pydantic==2.5.3
shapely==2.0.2
geopandas==0.14.2
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
